+++ nacc_attribute_deriver/attributes/nacc/modules/uds/__init__.py
@@ -1,4 +1,6 @@
 # ruff: noqa
 from .form_a1 import *
+from .form_a4 import *
+from .form_a5 import *
 from .form_d1 import *
 from .form_header import *