+++ nacc_attribute_deriver/attributes/nacc/modules/uds/form_a2.py
@@ -1,4 +1,7 @@
-"""Derived variables from form A2: Co-participant Demographics."""
+"""Derived variables from form A2: Co-participant Demographics.
+
+From a2structrdd.sas
+"""
 
 from typing import Optional